refactor: move Input and Output classes to 'io' package
